+++ rana_qgis_plugin/auth.py
@@ -1,8 +1,9 @@
 import json
 
-from qgis.core import QgsApplication, QgsAuthMethodConfig, QgsMessageLog
+from qgis.core import QgsApplication, QgsAuthMethodConfig
 from qgis.PyQt.QtCore import QSettings
 
+from .communication import UICommunication
 from .constant import (
     COGNITO_AUTHENTICATION_ENDPOINT,
     COGNITO_CLIENT_ID,
@@ -18,7 +19,7 @@ def get_authcfg_id():
     return authcfg_id
 
 
-def setup_oauth2():
+def setup_oauth2(communication: UICommunication):
     settings = QSettings()
     auth_manager = QgsApplication.authManager()
     auth_manager.setMasterPassword()
@@ -61,4 +62,4 @@ def setup_oauth2():
     if new_authcfg_id:
         settings.setValue(RANA_AUTHCFG_ENTRY, new_authcfg_id)
     else:
-        QgsMessageLog("Failed to create OAuth2 configuration", "Rana", QgsMessageLog.CRITICAL)
+        communication.log_warn("Failed to create OAuth2 configuration")